+++ projects/amdsmi/tests/amd_smi_test/main.cc
@@ -41,6 +41,7 @@
 #include "functional/gpu_partition_metrics_read.h"
 #include "functional/hw_topology_read.h"
 #include "functional/id_info_read.h"
+#include "functional/idle_metrics_read.h"
 #include "functional/kfd_atfork_read.h"
 #include "functional/mem_page_info_read.h"
 #include "functional/mem_util_read.h"
@@ -266,6 +267,10 @@ TEST(amdsmitstReadOnly, TestHWTopologyRead) {
 
 TEST(amdsmitstReadOnly, TestGpuMetricsRead) {
   TestGpuMetricsRead tst;
+  RunGenericTest(&tst);
+}
+TEST(amdsmitstReadOnly, TestIdleMetricsRead) {
+  TestIdleMetricsRead tst;
   RunGenericTest(&tst);
 }
 TEST(amdsmitstReadOnly, TestGpuPartitionMetricsRead) {